Additions to limelight commands as well as the restructuring of comamnds into subpackages
