+++ src/main/java/frc/robot/RobotContainer.java
@@ -25,12 +25,12 @@ import net.bancino.robotics.swerveio.exception.SwerveRuntimeException;
 import net.bancino.robotics.swerveio.command.SwerveDriveTeleop;
 import net.bancino.robotics.swerveio.command.PathweaverSwerveDrive;
 import net.bancino.robotics.swerveio.command.RunnableCommand;
-import frc.robot.commands.ElevatorWithJoystick;
-import frc.robot.commands.FeedWithJoystick;
-import frc.robot.commands.IntakeWithJoystick;
-import frc.robot.commands.LimelightAlignBackHatch;
-import frc.robot.commands.LimelightAlignFrontHatch;
-import frc.robot.commands.ShooterWithJoystick;
+import frc.robot.commands.joystick.ElevatorWithJoystick;
+import frc.robot.commands.joystick.FeedWithJoystick;
+import frc.robot.commands.joystick.IntakeWithJoystick;
+import frc.robot.commands.joystick.ShooterWithJoystick;
+import frc.robot.commands.vision.LimelightAlignBackHatch;
+import frc.robot.commands.vision.LimelightAlignFrontHatch;
 import net.bancino.robotics.swerveio.gyro.NavXGyro;
 import net.bancino.robotics.jlimelight.Limelight;
 import net.bancino.robotics.liboi.DeadbandedXboxController;
