+++ src/main/java/frc/robot/commands/joystick/FeedWithJoystick.java
@@ -5,37 +5,41 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.joystick;
 
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Shooter;
+import frc.robot.Const;
+import frc.robot.subsystems.Feed;
 
-public class ShooterWithJoystick extends CommandBase {
+public class FeedWithJoystick extends CommandBase {
 
-    private Shooter shooter;
-    private XboxController.Button shooterButton;
-    private XboxController.Axis hoodAxis;
+    private Feed feed;
+    private XboxController.Button runButton, reverseButton;
     private XboxController xbox;
 
-    /** Creates a new ShooterWithJoystick, of course. */
-    public ShooterWithJoystick(Shooter shooter, XboxController xbox, XboxController.Button shooterButton, XboxController.Axis hoodAxis) {
-        this.shooter = shooter;
+    public FeedWithJoystick(Feed feed, XboxController xbox, XboxController.Button runButton, XboxController.Button reverseButton) {
+        this.feed = feed;
         this.xbox = xbox;
-        this.shooterButton = shooterButton;
-        this.hoodAxis = hoodAxis;
-        addRequirements(shooter);
+        this.runButton = runButton;
+        this.reverseButton = reverseButton;
+
+        addRequirements(feed);
     }
 
     // Called every time the scheduler runs while the command is scheduled.
     @Override
     public void execute() {
-        if (xbox.getRawButton(shooterButton.value)) {
-            shooter.run();
+
+        if (xbox.getRawButton(runButton.value)) {
+            double speed = Const.Speed.FEED_SPEED;
+            if (xbox.getRawButton(reverseButton.value)) {
+                speed *= -1;
+            }
+            feed.runAt(speed);
         } else {
-            shooter.stop();
+            feed.stop();
         }
-        shooter.setHoodSpeed(xbox.getRawAxis(hoodAxis.value));
     }
 
     // Returns true when the command should end.
